[fix] 설명회 신청 인원 조회 수정, 설명회 신청 내역 수정, 북마크 배너 수정

+++ src/main/java/FIS/iLUVit/controller/UserController.java
@@ -46,5 +46,5 @@ public class UserController {
     }
 
     @DeleteMapping("/alarm")
-    public Long deleteAlarm(@Login Long userId, @RequestBody AlarmDeleteDto request){ return userService.deleteUserAlarm(userId, request.getAlarmId()); }
+    public Integer deleteAlarm(@Login Long userId, @RequestBody AlarmDeleteDto request){ return userService.deleteUserAlarm(userId, request.getAlarmId()); }
 }
